make rrdtool compile with gcc 4.5 if -std=c99 is set -- anicka@suse.cz
git-svn-id: svn://svn.oetiker.ch/rrdtool/branches/1.2/program@2069 a5681a0c-68f1-0310-ab6d-d61299d08faa

+++ src/rrd_format.h
@@ -22,7 +22,7 @@
 /* #define RRD_VERSION   "0002" */
 /* changed because microsecond precision requires another field */
 #define RRD_VERSION   "0003"
-#define FLOAT_COOKIE  8.642135E130
+#define FLOAT_COOKIE   ((double)8.642135E130)
 
 #include "rrd_nan_inf.h"
